+++ routers/StudentRouterds.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const Studentrouter = express.Router();
 const multer = require("multer");
-const { add, List, deleteAll, deleteitem } = require("../controllers/StudentControllers");
+const { add, List, deleteAll, deleteitem, addId, UserID } = require("../controllers/StudentControllers");
 
 const storage = multer.diskStorage({
     destination: (req, file, cb) => {  // <-- Use 'cb' instead of 'cd'
@@ -19,4 +19,6 @@ Studentrouter.post("/upload", uploads.single("file"), add);
 Studentrouter.get("/data", List);
 Studentrouter.post("/delete", deleteitem); // New delete route
 Studentrouter.delete("/delete-all", deleteAll); // New delete route
+Studentrouter.post("/notify", addId); // New delete route
+Studentrouter.post("/user", UserID); // New delete route
 module.exports = Studentrouter;
